add scalar to updateGraph and add updateBattery
added updateBattery function. This function takes the b array filled with advertised battery life data and adds it to the "mybattery" TextView, which will display it. The function only updates the View every 50 cycles to save computing work, and the conversion equation used to transform the value in the b array to a percent was found experimentally and may be slightly off. Additionally, the .0035 scalar was added to each array for updateGraph() to convert the values to flow rates.