Fixed ViewPager
Adjusted a bug with the ViewPager where it would create new instances of fragments instead of accessing the original fragments by creating a list holding all the original fragments. Also the app starts the timer once the next button in instructionfragment is pressed. Finally, the timerButton is hidden until the timer is finished so users must wait before moving on.